multimodal-ocr-vlm-serving-stack/pyproject.toml
[project]
name = "multimodal-ocr-vlm-serving-stack"
version = "0.1.0"
description = "Queue-backed OCR and VLM serving stack with batching, memory-aware scheduling, and observability."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic>=2.8,<3.0",
  "pydantic-settings>=2.4,<3.0",
  "prometheus-client>=0.20,<1.0",
  "httpx>=0.27,<1.0",
  "python-multipart>=0.0.9",
  "Pillow>=10.4,<11.0",
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.23,<1.0"
]

[project.optional-dependencies]
redis = ["redis>=5.0,<6.0"]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
